Struct draw_state::state::Rasterizer

pub struct Rasterizer {
    pub front_face: FrontFace,
    pub cull_face: CullFace,
    pub method: RasterMethod,
    pub offset: Option<Offset>,
    pub samples: Option<MultiSample>,
}

Primitive rasterization state. Note that GL allows different raster method to be used for front and back, while this abstraction does not.

Fields
front_face: FrontFace

Which vertex winding is considered to be the front face for culling.

cull_face: CullFace

Which face should be culled.

method: RasterMethod

How to rasterize this primitive.

offset: Option<Offset>

Any polygon offset to apply.

samples: Option<MultiSample>

Multi-sampling mode.

Methods

impl Rasterizer

fn new_fill() -> Rasterizer

Create a new filling rasterizer.

fn with_cull_back(self) -> Rasterizer

Add back face culling.

fn with_offset(self, slope: f32, units: OffsetUnits) -> Rasterizer

Add polygon offset.